CI: Allow to skip clone step of craft

+++ setup/CraftBootstrap.py
@@ -134,11 +134,14 @@ def setUp(args):
     if args.set:
         writeSettings(args)
 
-    craftDir = os.path.join(args.root, "craft")
-    verbosityFlag = "-vvv" if args.verbose else ""
-    run(args, f"craft --ci-mode --no-cache {verbosityFlag} git")
-    run(args, f"git clone --branch={args.branch} kde:craft {craftDir}")
-    shutil.rmtree(os.path.join(args.root, f"craft-{args.branch}"))
+    if not args.noCloneCraft:
+        craftDir = os.path.join(args.root, "craft")
+        verbosityFlag = "-vvv" if args.verbose else ""
+        run(args, f"craft --ci-mode --no-cache {verbosityFlag} git")
+        run(args, f"git clone --branch={args.branch} kde:craft {craftDir}")
+        shutil.rmtree(os.path.join(args.root, f"craft-{args.branch}"))
+    else:
+        shutil.move(os.path.join(args.root, f"craft-{args.branch}"), os.path.join(args.root, "craft"))
     print("Setup complete")
     print(f"Please run {args.root}/craft/kdeenv.ps1")
 
@@ -169,6 +172,7 @@ if __name__ == "__main__":
     parser.add_argument("--architecture", action="store")
     parser.add_argument("--no-short-path", action="store_true", dest="noShortPath")
     parser.add_argument("--no-bootstrap", action="store_true", dest="noBootstrap")
+    parser.add_argument("--no-clone-craft", action="store_true", dest="noCloneCraft")
     parser.add_argument("--set", action="store_true")
     parser.add_argument("--verbose", action="store_true")
